[FEAT] 매칭 게시글 CRUD
로그인한 유저 등록, 날짜 테스트 필요

+++ src/main/java/com/hibit/kusitms26tht3hibitback/repository/MatchingRepository.java
@@ -4,6 +4,6 @@ import com.hibit.kusitms26tht3hibitback.domain.Matching;
 import com.hibit.kusitms26tht3hibitback.domain.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MatchingRepository extends JpaRepository<Matching, Long> {
+public interface MatchingRepository extends JpaRepository<Matching, Integer> {
 
 }
